+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.96.1
 
+- Node: Custom registration errors